model/requirements.txt
onnx==1.15.0
onnxruntime==1.16.3
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
pillow==10.2.0
jupyter==1.0.0